refactor(modules/items/tracks): move validating from `TracksService` to `TracksValidator`
