+++ src/modules/items/tracks/tracks.module.ts
@@ -5,6 +5,7 @@ import { Track } from './track.entity'
 import { TracksRepository } from './tracks.repository'
 import { TracksController } from './tracks.controller'
 import { TracksService } from './tracks.service'
+import { TracksValidator } from './tracks.validator'
 
 import { SpotifyModule } from '@modules/spotify'
 import { AlbumsModule } from '@modules/items/albums'
@@ -16,7 +17,7 @@ import { AlbumsModule } from '@modules/items/albums'
     forwardRef(() => AlbumsModule),
   ],
   controllers: [TracksController],
-  providers: [TracksRepository, TracksService],
+  providers: [TracksRepository, TracksService, TracksValidator],
   exports: [TracksRepository, TracksService],
 })
 export class TracksModule {}
